Hotfix frontend
Updated Dockerfile: When the application is started with the Debug flag, it reboots upon each update of files in the directory. However, upon restart, the application attempts to initiate from PYTHONPATH, which does not include our /app directory. This results in the 'main not found' error. This update addresses the issue.